Housekeeping work. Integration of new i18n.jar release.
git-svn-id: http://argouml.tigris.org/svn/argouml/trunk/src_new@5230 a161b567-7d1e-0410-9ef9-912c70fedb3f

+++ org/argouml/xml/todo/TodoParser.java
@@ -526,7 +526,7 @@ public class TodoParser extends SAXParserBase {
 	    {
 		if (i2 > i1)
 		    sb.append(str.substring(i1, i2));
-		sb.append("%" + Integer.toString((int) c) + ";");
+		sb.append("%" + Integer.toString(c) + ";");
 		i1 = i2 + 1;
 	    }
 	}
